requirements.txt
python-docx>=1.0.0
google-generativeai>=0.3.0
pytest>=7.1.3
black>=22.10.0
pylint>=2.15.5
pytest-cov>=4.0.0
